chore: move storage out of internal
